Restart mypy daemon when settings change
This uses the fact that dmypy will restart itself if a plugin changes

Either by the `__version__` where the plugin lives changing, or the
module the plugin is in changing

We use the fact that we can change this value before it's read to change
the version if installed_apps changes. We use a consistent hash of the
settings.INSTALLED_APPS value to determine this hash.

We also get the setting by spawning a new process to avoid problems in
dmypy with the import space existing