Fix Invalid cross-device link
Prior to this when the reports are being created we protect against
partial writes by creating the files in a temporary location and then
renaming those files.

It turns out that it's very possible the default temporary location and
the scratch path are on different devices and the pathlib rename
functionality doesn't know how to deal with that.

This change fixes that by making sure the temporary location is added to
the scratch path so that it's on the same device