Cleanup type_checker code to resolve annotations more thoroughly
Prior to this the plugin hooks were too eager and also we were running
into the possibility of trying to resolve variables in the wrong scope.

This also cleans up the type_checker action in general.

The next commit will fix a bug that is introduced here where
annotations based on type variables aren't expanded after the type
variable is resolved